changed framework to three-d

+++ src/config.rs
@@ -5,11 +5,11 @@
 pub mod default {
 
     //pub const WINDOW_S: Vec2 = Vec2::new(0., 0.);
-    pub const WINDOW_S: (f32, f32) = (1200.0, 900.0);
+    pub const WINDOW_S: (u32, u32) = (1200, 900);
 }
 
 pub mod gui {
-    use bevy_egui::egui;
+    use three_d::egui;
 
     pub const fn shaded_color(darkmode: bool) -> egui::Color32 {
         match darkmode {
